development
minor improvements to ui

bugfixes related to recent changes to billing view model handling

refactored printviewmodel and moved most of the code to helper classes

further implementation of FTI duration, print and billing view model

moved tenant assigenment to its own viewmodel and view
