+++ SharedLivingCostCalculator.csproj.user
@@ -2,170 +2,12 @@
 <Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup />
   <ItemGroup>
-    <ApplicationDefinition Update="App.xaml">
-      <SubType>Designer</SubType>
-    </ApplicationDefinition>
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Update="Components\CurrencyInputView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Components\DoubleInputView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Components\FlatInfo.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Components\ObjectInputView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Manual\ManualPages\FlatManagementPage.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Manual\ManualPages\MainPage.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Manual\ManualPages\SettingsPage.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\AccountingView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\AnnualBillingView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\ConsumptionView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\CreditSetupRentView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\CreditSetupView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\FlatSetupView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\OtherCostsRentView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\OtherCostsSetupView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\OtherCostsView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\PrintView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\RoomSetupView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\SettingsView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\TenantSetupView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\FlatManagementView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\PaymentManagementView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\PaymentsSetupView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\RentManagementView.xaml.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Update="Views\RentUpdateView.xaml.cs">
+    <Compile Update="Views\TenantAssignmentView.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <Page Update="Components\CurrencyInputView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Components\DoubleInputView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Components\FlatInfo.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Components\ObjectInputView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="MainWindow.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Manual\ManualPages\FlatManagementPage.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Manual\ManualPages\MainPage.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Manual\ManualPages\SettingsPage.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Themes\Default.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Themes\Generic.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\AccountingView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\AnnualBillingView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\ConsumptionView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\CreditSetupRentView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\CreditSetupView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\FlatSetupView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\OtherCostsRentView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\OtherCostsSetupView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\OtherCostsView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\PrintView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\RoomSetupView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\SettingsView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\TenantSetupView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\FlatManagementView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\PaymentManagementView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\PaymentsSetupView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\RentManagementView.xaml">
-      <SubType>Designer</SubType>
-    </Page>
-    <Page Update="Views\RentUpdateView.xaml">
+    <Page Update="Views\TenantAssignmentView.xaml">
       <SubType>Designer</SubType>
     </Page>
   </ItemGroup>
