+++ Utility/PrintViewHelperFunctions/PrintOutputBase.cs
@@ -1,24 +1,158 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Documents;
+﻿using SharedLivingCostCalculator.ViewModels.Contract.ViewLess;
+using SharedLivingCostCalculator.ViewModels.Financial.ViewLess;
 using System.Windows;
+using System.Windows.Documents;
+using System.Collections.ObjectModel;
+using System.Windows.Media;
+using System.Data;
 
 namespace SharedLivingCostCalculator.Utility.PrintViewHelperFunctions
 {
     public class PrintOutputBase
     {
-        public Table GetOutputTableForFlat => OutputTableForFlat();
 
-        public PrintOutputBase()
+        private FlatViewModel _FlatViewModel;
+        public FlatViewModel FlatViewModel => _FlatViewModel;
+
+
+        public int SelectedYear {  get; set; }
+
+
+        public PrintOutputBase(FlatViewModel flatViewModel, int selectedYear)
         {
-            
+            _FlatViewModel = flatViewModel;
+            SelectedYear = selectedYear;
         }
 
 
-        private Table OutputTableForFlat()
+        public string BuildAddressDetails()
+        {
+            //Section section = new Section();
+
+            //string ReportHeader = "";
+            string rooms = "";
+
+            try
+            {
+                //ReportHeader = Application.Current.FindResource("IDF_Address").ToString();
+                rooms = Application.Current.Resources["IDF_Rooms"].ToString();
+            }
+            catch (Exception)
+            {
+            }
+
+            //Paragraph p = new Paragraph(new Run(ReportHeader));
+            //p.Style = headerParagraph;
+            //section.Blocks.Add(p);
+
+            //p = new Paragraph(new Run(FlatViewModel.Address));
+            //p.Style = textParagraph;
+            //section.Blocks.Add(p);
+
+            //p = new Paragraph(new Run($"{_FlatViewModel.Area}m², {_FlatViewModel.RoomCount} {rooms}"));
+            //p.Style = textParagraph;
+            //section.Blocks.Add(p);
+
+            return $"{FlatViewModel.Address}, {_FlatViewModel.Area}m², {_FlatViewModel.RoomCount} {rooms}";
+        }
+
+
+
+        public ObservableCollection<RentViewModel> FindRelevantRentViewModels()
+        {
+            ObservableCollection<RentViewModel> preSortList = new ObservableCollection<RentViewModel>();
+            ObservableCollection<RentViewModel> RentList = new ObservableCollection<RentViewModel>();
+
+            DateTime startDate = new DateTime(SelectedYear, 1, 1);
+            DateTime endDate = new DateTime(SelectedYear, 12, 31);
+
+            if (FlatViewModel.RentUpdates.Count > 0)
+            {
+                // filling the collection with potential matches
+                foreach (RentViewModel rent in FlatViewModel.RentUpdates)
+                {
+                    // rent begins after selected year ends
+                    if (rent.StartDate.Year > SelectedYear)
+                    {
+                        continue;
+                    }
+
+                    // rent begins before selected year starts
+                    if (rent.StartDate < startDate)
+                    {
+                        preSortList.Add(new RentViewModel(FlatViewModel, rent.Rent));
+                        continue;
+                    }
+
+                    // rent begins before selected year ends
+                    if (rent.StartDate < endDate)
+                    {
+                        preSortList.Add(new RentViewModel(FlatViewModel, rent.Rent));
+
+                        continue;
+                    }
+
+                    // rent begins after Billing period start but before Billing period end
+                    if (rent.StartDate > startDate || rent.StartDate < endDate)
+                    {
+                        preSortList.Add(new RentViewModel(FlatViewModel, rent.Rent));
+                    }
+                }
+
+                //    RentViewModel? comparer = new RentViewModel(_flatViewModel, new Rent() { StartDate = StartDate });
+                //    bool firstRun = true;
+
+                //    // building a collection of relevant rent items
+                //    foreach (RentViewModel item in preSortList)
+                //    {
+                //        if (item.StartDate >= StartDate)
+                //        {
+                //            RentList.Add(item);
+                //            continue;
+                //        }
+
+                //        if (item.StartDate < StartDate && firstRun)
+                //        {
+                //            firstRun = false;
+                //            comparer = item;
+                //            continue;
+                //        }
+
+                //        if (item.StartDate < StartDate && item.StartDate > comparer.StartDate)
+                //        {
+                //            comparer = item;
+                //        }
+                //    }
+                //    RentList.Add(comparer);
+            }
+
+            // sort List by StartDate, ascending
+            RentList = new ObservableCollection<RentViewModel>(preSortList.OrderBy(i => i.StartDate));
+
+            return RentList;
+        }
+
+
+        public Table OutputTableForFlatBilling()
+        {
+
+            Table dataOutputTable = new Table();
+
+            TableColumn StartDateColumn = new TableColumn() { Width = new GridLength(100) };
+            TableColumn EndDateColumn = new TableColumn() { Width = new GridLength(100) };
+            TableColumn ItemColumn = new TableColumn() { Width = new GridLength(250) };
+            TableColumn CostColumn = new TableColumn() { Width = new GridLength(100) };
+
+            dataOutputTable.Columns.Add(StartDateColumn);
+            dataOutputTable.Columns.Add(EndDateColumn);
+            dataOutputTable.Columns.Add(ItemColumn);
+            dataOutputTable.Columns.Add(CostColumn);
+
+            return dataOutputTable;
+        }
+
+
+        public Table OutputTableForFlat()
         {
 
             Table dataOutputTable = new Table();
@@ -33,5 +167,215 @@ namespace SharedLivingCostCalculator.Utility.PrintViewHelperFunctions
 
             return dataOutputTable;
         }
+
+
+        public Table OutputTableRooms()
+        {
+
+            Table dataOutputTable = new Table();
+
+            TableColumn RoomNameColumn = new TableColumn() { Width = new GridLength(150) };
+            TableColumn DateColumn = new TableColumn() { Width = new GridLength(100) };
+            TableColumn ItemColumn = new TableColumn() { Width = new GridLength(250) };
+            TableColumn CostColumn = new TableColumn() { Width = new GridLength(100) };
+
+            dataOutputTable.Columns.Add(RoomNameColumn);
+            dataOutputTable.Columns.Add(DateColumn);
+            dataOutputTable.Columns.Add(ItemColumn);
+            dataOutputTable.Columns.Add(CostColumn);
+
+            return dataOutputTable;
+        }
+
+
+        public TableRow OutputTableRow(RentViewModel viewModel, double payment, string item, int month = -1, bool FontWeightBold = false)
+        {
+            TableRow dataRow = new TableRow();
+
+            TableCell DueTime = new TableCell();
+            DueTime.TextAlignment = TextAlignment.Right;
+
+            TableCell Item = new TableCell();
+            TableCell Payment = new TableCell();
+            Payment.TextAlignment = TextAlignment.Right;
+
+            if (month == -1 && viewModel.StartDate.Year == SelectedYear)
+            {
+                DueTime.Blocks.Add(new Paragraph(new Run($"> {viewModel.StartDate.Month}/{SelectedYear}")) { Margin = new Thickness(0, 0, 10, 0) });
+            }
+            else if (month == -1 && viewModel.StartDate.Year < SelectedYear)
+            {
+                DueTime.Blocks.Add(new Paragraph(new Run($"> 1/{SelectedYear}")) { Margin = new Thickness(0, 0, 10, 0) });
+            }
+            else
+            {
+                DueTime.Blocks.Add(new Paragraph(new Run($"{month}/{SelectedYear}")) { Margin = new Thickness(0, 0, 10, 0) });
+            }
+
+            Item.Blocks.Add(new Paragraph(new Run(item)));
+
+            if (FontWeightBold)
+            {
+                Paragraph paymentParagraph = new Paragraph(new Run($"{payment:C2}"))
+                {
+                    BorderBrush = new SolidColorBrush(Colors.Black),
+                    BorderThickness = new Thickness(0, 1, 0, 0),
+                    FontWeight = FontWeights.Bold,
+                    Margin = new Thickness(0, 0, 10, 0)
+                };
+
+                Payment.Blocks.Add(paymentParagraph);
+            }
+            else
+            {
+                Paragraph paymentParagraph = new Paragraph(new Run($"{payment:C2}"))
+                {
+                    Margin = new Thickness(0, 0, 10, 0)
+                };
+
+                Payment.Blocks.Add(paymentParagraph);
+            }
+
+            dataRow.Cells.Add(DueTime);
+            dataRow.Cells.Add(Item);
+            dataRow.Cells.Add(Payment);
+
+            return dataRow;
+        }
+
+
+        public TableRow OutputTableRowRooms(RentViewModel viewModel, string roomname, double payment, string item, int month = -1, bool FontWeightBold = false)
+        {
+            TableRow dataRow = new TableRow();
+
+            TableCell RoomName = new TableCell();
+
+            TableCell DueTime = new TableCell();
+            DueTime.TextAlignment = TextAlignment.Right;
+
+            TableCell Item = new TableCell();
+            TableCell Payment = new TableCell();
+            Payment.TextAlignment = TextAlignment.Right;
+
+            RoomName.Blocks.Add(new Paragraph(new Run(roomname)));
+
+            if (month == -1 && viewModel.StartDate.Year == SelectedYear)
+            {
+                DueTime.Blocks.Add(new Paragraph(new Run($"> {viewModel.StartDate.Month}/{SelectedYear}")) { Margin = new Thickness(0, 0, 10, 0) });
+            }
+            else if (month == -1 && viewModel.StartDate.Year < SelectedYear)
+            {
+                DueTime.Blocks.Add(new Paragraph(new Run($"> 1/{SelectedYear}")) { Margin = new Thickness(0, 0, 10, 0) });
+            }
+            else
+            {
+                DueTime.Blocks.Add(new Paragraph(new Run($"{month}/{SelectedYear}")) { Margin = new Thickness(0, 0, 10, 0) });
+            }
+
+            Item.Blocks.Add(new Paragraph(new Run(item)));
+
+            if (FontWeightBold)
+            {
+                Paragraph paymentParagraph = new Paragraph(new Run($"{payment:C2}\n"))
+                {
+                    BorderBrush = new SolidColorBrush(Colors.Black),
+                    BorderThickness = new Thickness(0, 1, 0, 0),
+                    FontWeight = FontWeights.Bold,
+                    Margin = new Thickness(0, 0, 10, 0)
+                };
+
+                Payment.Blocks.Add(paymentParagraph);
+            }
+            else
+            {
+                Paragraph paymentParagraph = new Paragraph(new Run($"{payment:C2}"))
+                {
+                    Margin = new Thickness(0, 0, 10, 0)
+                };
+
+                Payment.Blocks.Add(paymentParagraph);
+            }
+
+            dataRow.Cells.Add(RoomName);
+            dataRow.Cells.Add(DueTime);
+            dataRow.Cells.Add(Item);
+            dataRow.Cells.Add(Payment);
+
+            return dataRow;
+        }
+
+
+        public BillingViewModel? SearchForBillingViewModel()
+        {
+            foreach (BillingViewModel billingViewModel in FlatViewModel.AnnualBillings)
+            {
+                if (billingViewModel.Year == SelectedYear)
+                {
+                   return billingViewModel;                    
+                }
+            }
+
+            return null;
+        }
+
+
+        public TableRowGroup TableRowGroupRentHeader()
+        {
+            TableRowGroup headerRowGroup = new TableRowGroup();
+
+            headerRowGroup.Style = Application.Current.FindResource("HeaderRowStyle") as Style;
+
+            headerRowGroup.FontSize = 14;
+
+            TableRow headerRow = new TableRow();
+
+            TableCell headerCell_DueTime = new TableCell();
+            TableCell headerCell_Item = new TableCell();
+            TableCell headerCell_Costs = new TableCell();
+
+            headerCell_DueTime.Blocks.Add(new Paragraph(new Run("Time")));
+            headerCell_Item.Blocks.Add(new Paragraph(new Run("Item")));
+            headerCell_Costs.Blocks.Add(new Paragraph(new Run("Costs")));
+
+            headerRow.Cells.Add(headerCell_DueTime);
+            headerRow.Cells.Add(headerCell_Item);
+            headerRow.Cells.Add(headerCell_Costs);
+
+            headerRowGroup.Rows.Add(headerRow);
+
+            return headerRowGroup;
+        }
+
+
+        public TableRowGroup TableRowGroupRoomHeader()
+        {
+            TableRowGroup headerRowGroup = new TableRowGroup();
+
+            headerRowGroup.Style = Application.Current.FindResource("HeaderRowStyle") as Style;
+
+            headerRowGroup.FontSize = 14;
+
+            TableRow headerRow = new TableRow();
+
+            TableCell RoomName = new TableCell();
+            TableCell DueTime = new TableCell();
+            TableCell Item = new TableCell();
+            TableCell Costs = new TableCell();
+
+            RoomName.Blocks.Add(new Paragraph(new Run("Room")));
+            DueTime.Blocks.Add(new Paragraph(new Run("Time")));
+            Item.Blocks.Add(new Paragraph(new Run("Item")));
+            Costs.Blocks.Add(new Paragraph(new Run("Costs")));
+
+            headerRow.Cells.Add(RoomName);
+            headerRow.Cells.Add(DueTime);
+            headerRow.Cells.Add(Item);
+            headerRow.Cells.Add(Costs);
+
+            headerRowGroup.Rows.Add(headerRow);
+
+            return headerRowGroup;
+        }
+
     }
 }
