+++ ViewModels/Financial/RentManagementViewModel.cs
@@ -192,8 +192,20 @@ namespace SharedLivingCostCalculator.ViewModels.Financial
 
             TransactionShareTypesRent shareType = item.TransactionShareTypes;
 
+            TransactionDurationTypes durationTypes = item.Duration;
+
+            DateTime endDate = new DateTime(item.EndDate.Year, item.EndDate.Month, item.EndDate.Day);
+            DateTime startDate = new DateTime(item.StartDate.Year, item.StartDate.Month, item.StartDate.Day);
+
             return new FinancialTransactionItemRentViewModel(
-                new FinancialTransactionItemRent() { TransactionItem = cause, TransactionSum = sum, TransactionShareTypes = shareType });
+                new FinancialTransactionItemRent() {
+                    TransactionItem = cause,
+                    TransactionSum = sum,
+                    TransactionShareTypes = shareType,
+                    StartDate = startDate,
+                    EndDate = endDate,
+                    Duration = durationTypes
+                });
         }
