+++ bin/Debug/net8.0-windows/appdata/appdata.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ApplicationDataOnExit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
-  <Accounting_Shown>false</Accounting_Shown>
-  <FlatManagement_Shown>true</FlatManagement_Shown>
+  <Accounting_Shown>true</Accounting_Shown>
+  <FlatManagement_Shown>false</FlatManagement_Shown>
   <FlatViewModelSelectedIndex>0</FlatViewModelSelectedIndex>
   <Manual_Shown>false</Manual_Shown>
   <Settings_Shown>false</Settings_Shown>
