+++ bin/Debug/net8.0-windows/resources.xml
@@ -41,16 +41,16 @@
     <ScB>0.06124606</ScB>
   </C_Text_Header>
   <ButtonCornerRadius>
+    <TopLeft>6</TopLeft>
+    <TopRight>6</TopRight>
+    <BottomRight>6</BottomRight>
+    <BottomLeft>6</BottomLeft>
+  </ButtonCornerRadius>
+  <VisibilityFieldCornerRadius>
     <TopLeft>7</TopLeft>
     <TopRight>7</TopRight>
     <BottomRight>7</BottomRight>
     <BottomLeft>7</BottomLeft>
-  </ButtonCornerRadius>
-  <VisibilityFieldCornerRadius>
-    <TopLeft>50</TopLeft>
-    <TopRight>50</TopRight>
-    <BottomRight>50</BottomRight>
-    <BottomLeft>50</BottomLeft>
   </VisibilityFieldCornerRadius>
   <FontFamily>Verdana</FontFamily>
   <FS>14</FS>
